server/requirements.txt
fastapi
firebase-admin
pydantic
uvicorn
python-dotenv
pyjwt
bcrypt